assignment 1 and 2
Question-1: Number Range Checker 
Write a program that prompts the user to enter a number. Check whether the number falls within any of the following ranges:
0-10: "Low Range"
11-50: "Medium Range"
51-100: "High Range"
Outside of 0-100: "Out of Range"

Question-2:Leap Year Checker 
Write a program that asks the user to input a year and determines whether the year is a leap year or not. A leap year is either divisible by 4 but not by 100, or divisible by 400.